fix:Models

+++ aserradero/src/Models/Albaran.ts
@@ -2,11 +2,11 @@ import { LinAlbar } from "./LinAlbar"
 import { Cliente } from "./Cliente.ts";
 
 export interface Albaran{
-    id:BigInteger,
+    id:bigint,
     numero:string,
     fecha:string,
     ano:number,
-    cliente_id:BigInteger,
+    cliente_id:bigint,
     total:number
     linAlbar:Array<LinAlbar>,
     Cliente:Cliente
