+++ aserradero/src/Models/LinAlbar.ts
@@ -2,9 +2,9 @@ import {Producto} from './Producto.ts';
 import { Albaran } from './Albaran.ts';
 
 export interface LinAlbar{
-    id:BigInteger,
-    producto_id:BigInteger,
-    albaran_id:BigInteger,
+    id:bigint,
+    producto_id:bigint,
+    albaran_id:bigint,
     unidades:number,
     L:number,
     A:number,
